Restructure 'executor' package + add windows memory usage info
- retrieve process memory usage via process handle
