+++ executor/execute_windows.go
@@ -1,0 +1,66 @@
+//go:build windows
+// +build windows
+
+package executor
+
+// executeCommand on Windows contains some dark sorcery. On Windows, the `rusage` equivalent does not include
+// memory information. In order to get this info, we need the process `handle`, not just its PID. Process
+// handle can be obtained by using `OpenProcess` syscall, but that is a data race, as the process might have
+// already exited by the time our syscall returns. To do this, we rely on the fact that the stdlib does not
+// change the process handle until a successful `Wait`. And on Windows, as long as we hold the handle, we
+// have access to the process information. So we'll use reflection to get the value of the handle and do a
+// `DuplicateHandle“ syscall. With this duplicated handle, we'll be able to access all the info we need.
+// Additionally, the `DuplicateHandle` syscall will fail if we do anything wrong, so it will also act as a
+// validation layer.
+func (e *Executor) executeCommand(cmd *exec.Cmd) (string, execute.Usage, error) {
+
+	var stdout bytes.Buffer
+	cmd.Stdout = &stdout
+
+	// Execute the command and collect output.
+	start := time.Now()
+	err := cmd.Start()
+	if err != nil {
+		return "", execute.Usage{}, fmt.Errorf("could not start process: %w", err)
+	}
+
+	childHandle, err := process.ReadHandle(cmd)
+	if err != nil {
+		return "", execute.Usage{}, fmt.Errorf("could not get process handle: %w", err)
+	}
+
+	// Create a duplicate handle - only for me (current process), not inheritable.
+	var handle windows.handle
+	me := windows.CurrentProcess()
+	err = windows.DuplicateHandle(me, childHandle, me, &handle, windows.PROCESS_QUERY_INFORMATION, false, 0)
+	if err != nil {
+		return "", execute.Usage{}, fmt.Errorf("could not duplicate process handle: %w", err)
+	}
+	defer func() {
+		err := windows.CloseHandle(childHandle)
+		if err != nil {
+			e.log.Error().Err(err).Int("pid", cmd.Process.Pid).Msg("could not close handle")
+		}
+	}()
+
+	// Now we can safely wait for the child process to complete.
+	err = cmd.Wait()
+	if err != nil {
+		return "", execute.Usage{}, fmt.Errorf("could not wait on process: %w", err)
+	}
+
+	end := time.Now()
+
+	// Create usage information.
+	duration := end.Sub(start)
+	usage, err := process.GetUsage(cmd)
+	if err != nil {
+		return "", execute.Usage{}, fmt.Errorf("could not retrieve usage data: %w", err)
+	}
+
+	// Returned memor usage is in bytes, so convert it to kilobytes.
+	usage.MemoryMaxKB = usage / 1000
+	usage.WallClockTime = duration
+
+	return stdout.String(), usage, nil
+}
