--- executor/function.go
@@ -1,34 +1,0 @@
-package executor
-
-import (
-	"fmt"
-
-	"github.com/blocklessnetworking/b7s/models/execute"
-	"github.com/blocklessnetworking/b7s/models/response"
-)
-
-// Function will execute the Blockless function defined by the execution request.
-func (e *Executor) Function(requestID string, req execute.Request) (execute.Result, error) {
-
-	// Execute the function.
-	out, usage, err := e.execute(requestID, req)
-	if err != nil {
-
-		res := execute.Result{
-			Code:      response.CodeError,
-			RequestID: requestID,
-			Usage:     usage,
-		}
-
-		return res, fmt.Errorf("function execution failed: %w", err)
-	}
-
-	res := execute.Result{
-		Code:      response.CodeOK,
-		RequestID: requestID,
-		Result:    out,
-		Usage:     usage,
-	}
-
-	return res, nil
-}
