+++ executor/internal/process/memory.go
@@ -1,7 +1,7 @@
 //go:build !windows
 // +build !windows
 
-package executor
+package process
 
 import (
 	"os"
